feat(elements): rename API to createCustomElement
